I moved everything onto the Router in order to make it a little more compact
