+++ smurfs/src/components/App.js
@@ -1,7 +1,15 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { addSmurf, getSmurfs } from '../actions';
-import Smurf from './Smurf';
+import { getSmurfs } from '../actions';
+
+
+import Smurfs from './Smurfs';
+import addForm from './addForm';
+
+// router stuff
+import {BrowserRouter as Router, Route, Link } from 'react-router-dom';
+
+
 import './App.css';
 /*
  to wire this component up you're going to need a few things.
@@ -11,81 +19,24 @@ import './App.css';
  */
 class App extends Component {
   state = {
-    smurfs: [],
-    smurf: {
-      name: '',
-      age: '',
-      height: ''
-    }
+    smurfs: []
   }
-
-  componentDidMount() {
-    this.props.getSmurfs();
-    //calling creator to populate smurfs.
-  }
-
-  handleChanges = e => {
-    this.setState({
-      smurf: {
-        ...this.state.smurf,
-        [e.target.name]: e.target.value
-      }
-    });
-  };
-
-  addHandler = e => {
-    e.preventDefault();
-    this.props.addSmurf(this.state.smurf);
-    this.setState({smurf: {
-      name: '',
-      age: '',
-      height: ''
-    }})
-  }
-
 
   render() {
-    if (this.props.fetching) {
-      return <div> Grabbing Smurfs </div>
-    }
     return (
+      <Router>
       <div className="App">
-        {this.props.smurfs.map(smurf => <Smurf key={smurf.name} smurf={smurf} />)}
-        <h2>Add a new Smurf </h2>
-        <form onSubmit={this.addHandler}>
-          <input
-          type="text"
-          name="name"
-          placeholder="Name"
-          value={this.state.smurf.name}
-          onChange={this.handleChanges}>
-          </input>
-          <input
-          type="text"
-          name="age"
-          placeholder="Age"
-          value={this.state.smurf.age}
-          onChange={this.handleChanges}>
-          </input>
-          <input
-          type="text"
-          name="height"
-          placeholder="Height"
-          value={this.state.smurf.height}
-          onChange={this.handleChanges}>
-          </input>
+        <Link to="/addsmurf">Add a Smurf</Link>
+        <Link to="/smurfs"> View Smurf Village </Link>
+        {/* {this.props.smurfs.map(smurf => <Smurf key={smurf.name} smurf={smurf} />)} */}
+      </div> 
 
-          <button>Submit</button>
-          </form> 
-      </div>
+      <Route path="/addsmurf" component={addForm} />
+      <Route path="/smurfs" component={Smurfs} />
+      </Router>
     );
   }
 }
 
 
-const MapStateToProps = state => ({
-        smurfs: state.smurfs,
-        fetching: state.fetchingSmurfs
-});
-
-export default connect(MapStateToProps, { addSmurf, getSmurfs})(App);
+export default App;
